+++ Wire-iOS Tests/ColorSchemeControllerTests.swift
